// src/frontend/dbmaker-frontend/api/consolidated/models/SettingsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SystemSettings } from './SystemSettings';
export type SettingsResponse = {
    settings?: SystemSettings;
    success?: boolean;
    message?: string | null;
};